+++ test/pty_event_failure_test.c
@@ -5,8 +5,11 @@
 
 #if defined(_WIN32)
 #include <windows.h>
+static volatile LONG post_count;
 #else
+#include <stdatomic.h>
 #include <unistd.h>
+static _Atomic int post_count;
 #endif
 
 #include "../src/flutter_pty.h"
@@ -16,7 +19,21 @@ static bool reject_post(Dart_Port_DL port, Dart_CObject *message)
 {
     (void)port;
     (void)message;
+#if defined(_WIN32)
+    InterlockedIncrement(&post_count);
+#else
+    atomic_fetch_add_explicit(&post_count, 1, memory_order_relaxed);
+#endif
     return false;
+}
+
+static int posted_event_count(void)
+{
+#if defined(_WIN32)
+    return (int)InterlockedCompareExchange(&post_count, 0, 0);
+#else
+    return atomic_load_explicit(&post_count, memory_order_acquire);
+#endif
 }
 
 static void wait_milliseconds(unsigned int milliseconds)
@@ -82,5 +99,6 @@ int main(void)
         wait_milliseconds(10);
     }
     assert(stats_are_zero());
+    assert(posted_event_count() == 1);
     return 0;
 }